refactor: Gemini 호출 공통 뼈대를 GeminiHttpClient 로 흡수하고 OCR 추출기 네이밍 정렬
- GeminiHttpClient 도입 — RestClient·timeout·헤더·generateContent 호출 + GeminiRetry 적용 + 에러 분류 + extractText + result 파싱까지 한 곳에서 처리
- GeminiExtractionResponse·GeminiOcrResponse 를 GeminiGenerateContentResponse 하나로 통합 (urlContextMetadata 는 optional, OCR 흐름엔 null)
- GeminiProductExtractor·GeminiOcrExtractor 가 자기 Request/Result 만 알면 되도록 HTTP 호출 코드 50+ 줄/파일을 위임으로 축소
- OcrClient → OcrExtractor, GeminiOcrClient → GeminiOcrExtractor, analyzeImage(image) → extract(image) 로 네이밍 정렬 (ProductExtractor 와 짝 맞춤)
- OcrImage KDoc 의 GeminiOcrClient 참조도 함께 갱신
