+++ src/main/kotlin/com/depromeet/team3/ocr/service/OcrService.kt
@@ -7,8 +7,8 @@ import org.springframework.web.multipart.MultipartFile
 
 @Service
 class OcrService(
-    private val ocrClient: OcrClient,
+    private val ocrExtractor: OcrExtractor,
 ) {
     // 입력 검증(빈 파일 · MIME 타입)은 OcrImage.of 가 담당한다. 서비스는 위임만 한다.
-    fun extract(image: MultipartFile): Product = ocrClient.analyzeImage(OcrImage.of(image.bytes, image.contentType))
+    fun extract(image: MultipartFile): Product = ocrExtractor.extract(OcrImage.of(image.bytes, image.contentType))
 }
